chore: bump start-sdk to 1.5.1 (1.22.2:1)
No upstream or behavioral changes. The 1.3.3 → 1.5.1 jump covers
several SDK-internal fixes (FileHelper.produce listener leak, zod 4.3.6
pin, GetActionInputType inference, BindOptions.addSsl type symmetry)
and additive surface (sdk.notification.create, FileHelper.yaml options,
ProxyAuth/BasicCredential, footnote/triState input specs) — none of
which the package currently touches, so no source changes are needed
beyond the dep bump.

Per the versions convention (no migration → rename in place, no new
file), rename v1.22.2.0.ts → v1.22.2.1.ts and refresh release notes
to a terse "Internal" bump across en_US/es_ES/de_DE/pl_PL/fr_FR. The
old-WORK_DIR migration is preserved and remains idempotent for users
coming from any earlier version.

npm update also refreshes the lnd-startos master pin in the lockfile.
