+++ startos/versions/index.ts
@@ -1,7 +1,7 @@
 import { VersionGraph } from '@start9labs/start-sdk'
-import { v_1_22_2_0 } from './v1.22.2.0'
+import { v_1_22_2_1 } from './v1.22.2.1'
 
 export const versionGraph = VersionGraph.of({
-  current: v_1_22_2_0,
+  current: v_1_22_2_1,
   other: [],
 })
